Added safe balls for each node
Every node has a ball around it which says that if the system starts
anywhere within the bowl and applies the control of that node, it will
definitely reach the bowl of the parent. Given the bowl of the root
state, these balls are currently propagated in a very adhoc manner.

To ponder -> this gives "a policy". However, it is sort of obvious that
as opposed to the full propagation, we are restraining ourselves only to
policies that follow the path of the node. Particles could take other
optimal paths if they strayed too far away from balls.

git-svn-id: svn+ssh://ares.lids.mit.edu/code/pratikac@1223 d6d04add-692a-0410-8435-c93ef0a91bfa